Release/gcp (#42)
* GCP relealse branhc

* Deployment ready
