+++ Inventory/.dockerignore
@@ -1,0 +1,6 @@
+.gradle
+.idea
+.vscode
+bin/
+gradle/
+temp/
