+++ Inventory/.gitignore
@@ -1,5 +1,6 @@
 .gradle
 build/
+config/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
