+++ Inventory/build.gradle
@@ -2,9 +2,7 @@ plugins {
 	id 'java'
 	id 'org.springframework.boot' version '3.0.0'
 	id 'io.spring.dependency-management' version '1.1.0'
-	id "com.diffplug.spotless" version "6.9.0"
 	id 'com.github.sherter.google-java-format' version '0.9'
-	id "org.flywaydb.flyway" version "9.8.1"
 }
 
 group = 'com.touchblankspot'
@@ -29,7 +27,6 @@ repositories {
 }
 
 dependencies {
-	implementation project('common')
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-security'
@@ -47,6 +44,8 @@ dependencies {
 	implementation 'commons-collections:commons-collections:3.2.2'
 	implementation 'org.apache.poi:poi:5.2.3'
 	implementation 'org.apache.poi:poi-ooxml:5.2.3'
+	implementation 'org.passay:passay:1.6.2'
+	implementation 'commons-beanutils:commons-beanutils:1.9.4'
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	runtimeOnly 'com.mysql:mysql-connector-j'
@@ -67,11 +66,11 @@ tasks.named('test') {
 }
 
 
-task prefixNewMigrations {
-	fileTree(dir:'src/main/resources/db/migration').exclude({ isFileOldPrefix(it.file) || isFilePrefixed(it.file) }).each {file ->
+tasks.register('prefixNewMigrations') {
+	fileTree(dir: 'src/main/resources/db/migration').exclude({ isFileOldPrefix(it.file) || isFilePrefixed(it.file) }).each { file ->
 		doLast {
 			def timestamp = new Date().format('yyyyMMdddHHmmssSSSS', TimeZone.getTimeZone('GMT'))
-			def name = "${file.name}".replaceAll('(?:V1)?_\\d+__','')
+			def name = "${file.name}".replaceAll('(?:V1)?_\\d+__', '')
 			file.renameTo("$file.parentFile.absolutePath$file.separator${'V2_' + timestamp}__${name}")
 			sleep(1000)
 		}
@@ -90,10 +89,6 @@ static def isFileOldPrefix(file){
 	return file.name.startsWith("V1__")
 }
 
-spotless {
-	java {
-		importOrder()
-		removeUnusedImports()
-		googleJavaFormat() // has its own section below
-	}
+jar{
+	include("/resource/**")
 }
