+++ Inventory/docker-compose.yml
@@ -1,0 +1,14 @@
+version: '3.3'
+
+services:
+  inventory:
+    container_name: inventory
+    build: .
+    image: inventory
+    ports:
+      - "8086:8086"
+    #networks:
+    #  - inventory_network
+#networks:
+#  inventory_network:
+#    external: true
