+++ Inventory/docker-compose_db.yml
@@ -1,0 +1,24 @@
+version: '3.3'
+services:
+  mysql_inventory:
+    container_name: mysql8_inventory
+    image: mysql:8
+    ports:
+      - "3308:3306"
+    restart: always
+    #volumes:
+    #  - ./mysql_data:/var/lib/mysql
+    networks:
+      - inventory_network
+    environment:
+      - MYSQL_HOST=localhost
+      - MYSQL_ROOT_PASSWORD=Pikr@1991
+      - TZ=Asia/Calcutta
+    healthcheck:
+      test: ["CMD", "mysqladmin" ,"ping", "-h", "localhost"]
+      timeout: 20s
+      retries: 10
+
+networks:
+  inventory_network:
+    external: true
