+++ Inventory/docker-execute.sh
@@ -1,0 +1,5 @@
+rm -rf build/libs
+./gradlew clean assemble
+docker-compose build
+docker-compose down
+docker-compose up -d
