+++ Inventory/init.sh
@@ -1,0 +1,2 @@
+docker network create --driver bridge inventory_network || true
+docker-compose -f docker-compose_db.yml up -d
