+++ Inventory/src/main/java/com/touchblankspot/common/Constant.java
