+++ Inventory/src/main/java/com/touchblankspot/common/validator/FieldValueExists.java
@@ -1,0 +1,6 @@
+package com.touchblankspot.common.validator;
+
+public interface FieldValueExists {
+
+  boolean fieldValueExists(Object value, String fieldName) throws UnsupportedOperationException;
+}
