+++ Inventory/src/main/java/com/touchblankspot/common/validator/IsUniqueInputFields.java
@@ -1,0 +1,9 @@
+package com.touchblankspot.common.validator;
+
+import java.util.List;
+
+public interface IsUniqueInputFields {
+
+  boolean isUniqueRowCombination(List<String> fields, List<Object> values, Boolean isUpdate)
+      throws UnsupportedOperationException;
+}
