+++ Inventory/src/main/java/com/touchblankspot/common/validator/PasswordConstraintValidator.java
@@ -34,7 +34,6 @@ public class PasswordConstraintValidator implements ConstraintValidator<ValidPas
   @SneakyThrows
   @Override
   public boolean isValid(String password, ConstraintValidatorContext context) {
-    // customizing validation messages
     Properties props = new Properties();
     InputStream inputStream = passyResource.getInputStream();
     props.load(inputStream);
@@ -43,23 +42,13 @@ public class PasswordConstraintValidator implements ConstraintValidator<ValidPas
         new PasswordValidator(
             resolver,
             Arrays.asList(
-                // length between 8 and 16 characters
                 new LengthRule(8, 16),
-                // at least one upper-case character
                 new CharacterRule(EnglishCharacterData.UpperCase, 1),
-                // at least one lower-case character
                 new CharacterRule(EnglishCharacterData.LowerCase, 1),
-                // at least one digit character
                 new CharacterRule(EnglishCharacterData.Digit, 1),
-                // at least one symbol (special character)
                 new CharacterRule(EnglishCharacterData.Special, 1),
-                // no whitespace
                 new WhitespaceRule(),
-                // rejects passwords that contain a sequence of >= 5 characters alphabetical  (e.g.
-                // abcdef)
                 new IllegalSequenceRule(EnglishSequenceData.Alphabetical, 5, false),
-                // rejects passwords that contain a sequence of >= 5 characters numerical   (e.g.
-                // 12345)
                 new IllegalSequenceRule(EnglishSequenceData.Numerical, 5, false)));
     RuleResult result = validator.validate(new PasswordData(password));
     if (result.isValid()) {
