+++ Inventory/src/main/java/com/touchblankspot/inventory/portal/GlobalExceptionHandler.java
@@ -20,15 +20,10 @@ public class GlobalExceptionHandler {
 
   @ExceptionHandler(value = Exception.class)
   public ModelAndView defaultErrorHandler(HttpServletRequest req, Exception e) throws Exception {
-    // If the exception is annotated with @ResponseStatus rethrow it and let
-    // the framework handle it - like the OrderNotFoundException example
-    // at the start of this post.
-    // AnnotationUtils is a Spring Framework utility class.
     if (AnnotationUtils.findAnnotation(e.getClass(), ResponseStatus.class) != null) {
       throw e;
     }
 
-    // Otherwise setup and send the user to a default error-view.
     ModelAndView mav = new ModelAndView();
     mav.addObject("exception", e);
     mav.addObject("url", req.getRequestURL());
@@ -38,15 +33,10 @@ public class GlobalExceptionHandler {
 
   @ExceptionHandler(value = AccessDeniedException.class)
   public ModelAndView accessDeniedHandler(HttpServletRequest req, Exception e) throws Exception {
-    // If the exception is annotated with @ResponseStatus rethrow it and let
-    // the framework handle it - like the OrderNotFoundException example
-    // at the start of this post.
-    // AnnotationUtils is a Spring Framework utility class.
     if (AnnotationUtils.findAnnotation(e.getClass(), ResponseStatus.class) != null) {
       throw e;
     }
 
-    // Otherwise setup and send the user to a default error-view.
     ModelAndView mav = new ModelAndView();
     mav.addObject("exception", e);
     mav.addObject("url", req.getRequestURL());
